+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Added workaround for https://github.com/aws/opsworks-cookbooks/issues/213 that will check to ensure bind mounts have been made before the deploy directory is created.
 # v3425 2015-07-27
 - ECS support for Amazon Linux and Ubuntu.
 # v3422 2015-06-29